pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Mobile Full-Stack Backend Service"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.2.1",
]
